requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
gitdb==4.0.10
GitPython==3.1.36
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.16.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.5.2
smmap==5.0.1
urllib3==2.0.4
